move window update tests to AbstractNettyTest
